requirements.txt
protobuf
streamlit
pyyaml
torch
transformers
scikit-learn
pandas
numpy
tqdm
datasets
pyarrow
kagglehub
